refactor(codegen): extract expression type queries
